change db to sqlite
