+++ model/sqlite.go
@@ -1,0 +1,23 @@
+package model
+
+import (
+	"gorm.io/driver/sqlite"
+	"gorm.io/gorm"
+	"time"
+)
+
+var DB *gorm.DB
+
+func InitDB() error {
+	var err error
+	DB, err = gorm.Open(sqlite.Open("data.db"), &gorm.Config{
+		NowFunc: func() time.Time {
+			return time.Now().UTC()
+		},
+	})
+	if err != nil {
+		return err
+	}
+	DB.AutoMigrate(&Client{})
+	return nil
+}
